add new environmental datasets in layer library
