+++ apps/fishing-map/data/workspaces.ts
@@ -56,6 +56,8 @@ export const TEMPLATE_VESSEL_DATAVIEW_SLUG = 'fishing-map-vessel-track'
 export const TEMPLATE_CONTEXT_DATAVIEW_SLUG = 'default-context-layer'
 export const TEMPLATE_POINTS_DATAVIEW_SLUG = 'default-points-layer'
 export const TEMPLATE_ENVIRONMENT_DATAVIEW_SLUG = 'default-environmental-layer'
+export const TEMPLATE_GFW_ENVIRONMENT_DATAVIEW_SLUG = 'gfw-environmental-layer'
+export const TEMPLATE_HEATMAP_ENVIRONMENT_DATAVIEW_SLUG = 'heatmap-environmental-layer'
 export const TEMPLATE_ACTIVITY_DATAVIEW_SLUG = 'activity-template'
 export const TEMPLATE_CLUSTERS_DATAVIEW_SLUG = 'template-for-bigquery-cluster-events'
 
@@ -64,6 +66,8 @@ export const TEMPLATE_DATAVIEW_SLUGS = [
   TEMPLATE_VESSEL_DATAVIEW_SLUG,
   TEMPLATE_CONTEXT_DATAVIEW_SLUG,
   TEMPLATE_ENVIRONMENT_DATAVIEW_SLUG,
+  TEMPLATE_GFW_ENVIRONMENT_DATAVIEW_SLUG,
+  TEMPLATE_HEATMAP_ENVIRONMENT_DATAVIEW_SLUG,
   TEMPLATE_POINTS_DATAVIEW_SLUG,
   TEMPLATE_ACTIVITY_DATAVIEW_SLUG,
   TEMPLATE_CLUSTERS_DATAVIEW_SLUG,
