+++ apps/fishing-map/features/layer-library/LayerLibrary.tsx
@@ -5,7 +5,7 @@ import { uniq } from 'lodash'
 import { useSelector } from 'react-redux'
 import { InputText } from '@globalfishingwatch/ui-components'
 import { DataviewCategory } from '@globalfishingwatch/api-types'
-import { LIBRARY_LAYERS, LibraryLayer } from 'data/library-layers'
+import { LIBRARY_LAYERS, LibraryLayer } from 'data/layer-library'
 import { upperFirst } from 'utils/info'
 import { selectAllDataviews } from 'features/dataviews/dataviews.slice'
 import LayerLibraryItem from 'features/layer-library/LayerLibraryItem'
